backend/pyproject.toml
[tool.poetry]
name = "iot-security-backend"
version = "1.0.0"
description = "IoT Security Testing Platform - Unified Backend"
package-mode = false
authors = ["Shantou AI Lab <dev@lab.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.13.0"
asyncpg = "^0.30.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = {extras = ["hiredis"], version = "^5.0.1"}
fastapi-limiter = "^0.1.5"
celery = {extras = ["redis"], version = "^5.3.4"}
minio = "^7.2.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}
pillow = "^10.1.0"
python-multipart = "^0.0.6"
websockets = "^12.0"
python-docx = "^1.1.0"
matplotlib = "^3.8.0"
psycopg2-binary = "^2.9.11"
email-validator = "^2.3.0"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
httpx = "^0.25.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
